Change training and test set split by doing:
"randomly selecting 400 patients from each of the 50 hospitals, and thus had a size of 20,000 examples; the test dataset contained the remaining 160 patients from each hos- pital, totaling 8000 examples"
instead of: " we selected 50 hospitals whose patient count was over 600 and, from each of them, randomly sampled 560 patients to form the final dataset of 28,000 examples. This data was split into a training set of 20,000 examples and a test set of 8000 examples so that the training- test ratio was about 7:3. Splitting data in this ratio is a common practice in fitting machine learning models to make biomedical predictions"